+++ src/main/java/com/istz/type/StaticBeans.java
@@ -6,21 +6,21 @@ public class StaticBeans {
 	// ゲームマスタ
 	public class GachaClasses{
 		private String    gameCd;
-		private Short     gachaKindTimes;
+		private Integer     gachaKindTimes;
 		private GachaKind gachaClass;
 
-		public GachaClasses(String gameCd, Short gachaKindTimes, GachaKind gachaClass) {
+		public GachaClasses(String gameCd, Integer gachaKindTimes, GachaKind gachaClass) {
 			this.gameCd         = gameCd;
 			this.gachaKindTimes = gachaKindTimes;
 			this.gachaClass     = gachaClass;
 		}
 
 		public String     getGameCd         (){return this.gameCd;}
-		public Short    getGachaKindTimes (){return this.gachaKindTimes;}
+		public Integer    getGachaKindTimes (){return this.gachaKindTimes;}
 		public GachaKind  getGachaClass (){return this.gachaClass;}
 
 		public void setGameCd         (String gameCd)         {this.gameCd   = gameCd;}
-		public void setGachaKindTimes (Short gachaKindTimes){this.gachaKindTimes = gachaKindTimes;}
+		public void setGachaKindTimes (Integer gachaKindTimes){this.gachaKindTimes = gachaKindTimes;}
 		public void setGachaClass     (GachaKind gachaClass)  {this.gachaClass = gachaClass;}
 
 	}
